cloudflare: move regional logic to dedicated file
