+++ provider/cloudflare/cloudflare_test.go
@@ -21,13 +21,12 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"reflect"
 	"slices"
 	"sort"
 	"strings"
 	"testing"
 
-	cloudflare "github.com/cloudflare/cloudflare-go"
+	"github.com/cloudflare/cloudflare-go"
 	"github.com/maxatome/go-testdeep/td"
 	log "github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
@@ -2984,337 +2983,6 @@ func TestCloudflareListCustomHostnamesWithPagionation(t *testing.T) {
 		t.Errorf("should not fail - %v", chErr)
 	}
 	assert.Equal(t, len(chs), CustomHostnamesNumber)
-}
-
-func Test_getRegionKey(t *testing.T) {
-	type args struct {
-		endpoint         *endpoint.Endpoint
-		defaultRegionKey string
-	}
-	tests := []struct {
-		name string
-		args args
-		want string
-	}{
-		{
-			name: "no region key",
-			args: args{
-				endpoint: &endpoint.Endpoint{
-					RecordType: "A",
-				},
-				defaultRegionKey: "",
-			},
-			want: "",
-		},
-		{
-			name: "default region key",
-			args: args{
-				endpoint: &endpoint.Endpoint{
-					RecordType: "A",
-				},
-				defaultRegionKey: "us",
-			},
-			want: "us",
-		},
-		{
-			name: "endpoint with region key",
-			args: args{
-				endpoint: &endpoint.Endpoint{
-					RecordType: "A",
-					ProviderSpecific: endpoint.ProviderSpecific{
-						{
-							Name:  "external-dns.alpha.kubernetes.io/cloudflare-region-key",
-							Value: "eu",
-						},
-					},
-				},
-				defaultRegionKey: "us",
-			},
-			want: "eu",
-		},
-		{
-			name: "endpoint with empty region key",
-			args: args{
-				endpoint: &endpoint.Endpoint{
-					RecordType: "A",
-					ProviderSpecific: endpoint.ProviderSpecific{
-						{
-							Name:  "external-dns.alpha.kubernetes.io/cloudflare-region-key",
-							Value: "",
-						},
-					},
-				},
-				defaultRegionKey: "us",
-			},
-			want: "",
-		},
-		{
-			name: "unsupported record type",
-			args: args{
-				endpoint: &endpoint.Endpoint{
-					RecordType: "TXT",
-					ProviderSpecific: endpoint.ProviderSpecific{
-						{
-							Name:  "external-dns.alpha.kubernetes.io/cloudflare-region-key",
-							Value: "eu",
-						},
-					},
-				},
-				defaultRegionKey: "us",
-			},
-			want: "",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := getRegionKey(tt.args.endpoint, tt.args.defaultRegionKey); got != tt.want {
-				t.Errorf("getRegionKey() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_dataLocalizationRegionalHostnamesChanges(t *testing.T) {
-	cmpDataLocalizationRegionalHostnameChange := func(i, j DataLocalizationRegionalHostnameChange) int {
-		if i.Action == j.Action {
-			return 0
-		}
-		if i.Hostname < j.Hostname {
-			return -1
-		}
-		return 1
-	}
-	type args struct {
-		changes []*cloudFlareChange
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    []DataLocalizationRegionalHostnameChange
-		wantErr bool
-	}{
-		{
-			name: "empty input",
-			args: args{
-				changes: []*cloudFlareChange{},
-			},
-			want:    nil,
-			wantErr: false,
-		},
-		{
-			name: "changes without RegionalHostname",
-			args: args{
-				changes: []*cloudFlareChange{
-					{
-						Action: cloudFlareCreate,
-						ResourceRecord: cloudflare.DNSRecord{
-							Name: "example.com",
-						},
-						RegionalHostname: cloudflare.RegionalHostname{}, // Empty
-					},
-				},
-			},
-			want:    nil,
-			wantErr: false,
-		},
-		{
-			name: "change with empty RegionKey",
-			args: args{
-				changes: []*cloudFlareChange{
-					{
-						Action: cloudFlareCreate,
-						ResourceRecord: cloudflare.DNSRecord{
-							Name: "example.com",
-						},
-						RegionalHostname: cloudflare.RegionalHostname{
-							Hostname:  "example.com",
-							RegionKey: "", // Empty region key
-						},
-					},
-				},
-			},
-			wantErr: true,
-		},
-		{
-			name: "conflicting region keys",
-			args: args{
-				changes: []*cloudFlareChange{
-					{
-						Action: cloudFlareCreate,
-						RegionalHostname: cloudflare.RegionalHostname{
-							Hostname:  "example.com",
-							RegionKey: "eu",
-						},
-					},
-					{
-						Action: cloudFlareCreate,
-						RegionalHostname: cloudflare.RegionalHostname{
-							Hostname:  "example.com",
-							RegionKey: "us", // Different region key for same hostname
-						},
-					},
-				},
-			},
-			wantErr: true,
-		},
-		{
-			name: "update takes precedence over create & delete",
-			args: args{
-				changes: []*cloudFlareChange{
-					{
-						Action: cloudFlareCreate,
-						RegionalHostname: cloudflare.RegionalHostname{
-							Hostname:  "example.com",
-							RegionKey: "eu",
-						},
-					},
-					{
-						Action: cloudFlareUpdate,
-						RegionalHostname: cloudflare.RegionalHostname{
-							Hostname:  "example.com",
-							RegionKey: "eu",
-						},
-					},
-					{
-						Action: cloudFlareDelete,
-						RegionalHostname: cloudflare.RegionalHostname{
-							Hostname:  "example.com",
-							RegionKey: "eu",
-						},
-					},
-				},
-			},
-			want: []DataLocalizationRegionalHostnameChange{
-				{
-					Action: cloudFlareUpdate,
-					RegionalHostname: cloudflare.RegionalHostname{
-						Hostname:  "example.com",
-						RegionKey: "eu",
-					},
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "create after delete becomes update",
-			args: args{
-				changes: []*cloudFlareChange{
-					{
-						Action: cloudFlareDelete,
-						RegionalHostname: cloudflare.RegionalHostname{
-							Hostname:  "example.com",
-							RegionKey: "eu",
-						},
-					},
-					{
-						Action: cloudFlareCreate,
-						RegionalHostname: cloudflare.RegionalHostname{
-							Hostname:  "example.com",
-							RegionKey: "eu",
-						},
-					},
-				},
-			},
-			want: []DataLocalizationRegionalHostnameChange{
-				{
-					Action: cloudFlareUpdate,
-					RegionalHostname: cloudflare.RegionalHostname{
-						Hostname:  "example.com",
-						RegionKey: "eu",
-					},
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "consolidate mixed actions for different hostnames",
-			args: args{
-				changes: []*cloudFlareChange{
-					{
-						Action: cloudFlareCreate,
-						RegionalHostname: cloudflare.RegionalHostname{
-							Hostname:  "example1.com",
-							RegionKey: "eu",
-						},
-					},
-					{
-						Action: cloudFlareUpdate,
-						RegionalHostname: cloudflare.RegionalHostname{
-							Hostname:  "example2.com",
-							RegionKey: "us",
-						},
-					},
-					{
-						Action: cloudFlareDelete,
-						RegionalHostname: cloudflare.RegionalHostname{
-							Hostname:  "example3.com",
-							RegionKey: "ap",
-						},
-					},
-					// duplicated actions
-					{
-						Action: cloudFlareCreate,
-						RegionalHostname: cloudflare.RegionalHostname{
-							Hostname:  "example1.com",
-							RegionKey: "eu",
-						},
-					},
-					{
-						Action: cloudFlareUpdate,
-						RegionalHostname: cloudflare.RegionalHostname{
-							Hostname:  "example2.com",
-							RegionKey: "us",
-						},
-					},
-					{
-						Action: cloudFlareDelete,
-						RegionalHostname: cloudflare.RegionalHostname{
-							Hostname:  "example3.com",
-							RegionKey: "ap",
-						},
-					},
-				},
-			},
-			want: []DataLocalizationRegionalHostnameChange{
-				{
-					Action: cloudFlareCreate,
-					RegionalHostname: cloudflare.RegionalHostname{
-						Hostname:  "example1.com",
-						RegionKey: "eu",
-					},
-				},
-				{
-					Action: cloudFlareUpdate,
-					RegionalHostname: cloudflare.RegionalHostname{
-						Hostname:  "example2.com",
-						RegionKey: "us",
-					},
-				},
-				{
-					Action: cloudFlareDelete,
-					RegionalHostname: cloudflare.RegionalHostname{
-						Hostname:  "example3.com",
-						RegionKey: "ap",
-					},
-				},
-			},
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := dataLocalizationRegionalHostnamesChanges(tt.args.changes)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("dataLocalizationRegionalHostnamesChanges() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			slices.SortFunc(got, cmpDataLocalizationRegionalHostnameChange)
-			slices.SortFunc(tt.want, cmpDataLocalizationRegionalHostnameChange)
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("dataLocalizationRegionalHostnamesChanges() = %v, want %v", got, tt.want)
-			}
-		})
-	}
 }
 
 func TestZoneHasPaidPlan(t *testing.T) {
